% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasSize.R
\name{canvasSize}
\alias{canvasSize}
\title{Get predefined canvas sizes}
\usage{
canvasSize(canvas)
}
\arguments{
\item{canvas}{Can either be a character with one of the values "standard",
"wide", "A4", "xeqy","wide-screen" or it can be a list with elements
width and height with single values (unit is inches).}
}
\description{
Get predefined canvas sizes
}
