% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNames.R
\name{compareNames}
\alias{compareNames}
\title{Compare element names in lists}
\usage{
compareNames(..., keepNames = T, testEqual = F, debug = F)
}
\arguments{
\item{...}{objects which element names to compare}

\item{keepNames}{If TRUE, the original dataset names are used in reported
table. If not, generic x1, x2,... are used. The latter may be preferred
for readability.}

\item{testEqual}{Do you just want a TRUE/FALSE to whether the names of the
two objects are the same? Default is FALSE which means to return an
overview for interactive use.}

\item{debug}{If TRUE, browser is called to begin with.}
}
\description{
Useful interactive tool when merging or binding objects
together. It lists the union of element names and indicates in
which of the objects they are present. It does not compare
contents of the elements at all.
}
\seealso{
Other DataWrangling: \code{\link{NMaddColumns}},
  \code{\link{NMcompAllRes}}, \code{\link{NMorderColumns}},
  \code{\link{NMscanDataDF}}, \code{\link{NMscanData}},
  \code{\link{factor2}}, \code{\link{findCovs_df}},
  \code{\link{findCovs}}, \code{\link{findVars}},
  \code{\link{mergeCheck}}, \code{\link{newNames}},
  \code{\link{objInfo}}, \code{\link{rbindIntersect}},
  \code{\link{rbindUnion}}, \code{\link{roundDF}},
  \code{\link{roundResp0}}, \code{\link{selectVars}},
  \code{\link{signif2}}, \code{\link{stampObj}},
  \code{\link{tmpcol}}
}
\concept{DataWrangling}
