% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{Calculate coefficient of variation}
\usage{
CV(x, log = F)
}
\arguments{
\item{x}{The data}

\item{log}{If TRUE, the geometric coefficient of variation is
calculated. This is sqrt(exp(var(log(x))-1).}
}
\description{
Calculate coefficient of variation
}
