% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLegend.R
\name{extractLegend}
\alias{extractLegend}
\title{Extract the legend from a ggplot}
\usage{
extractLegend(plot)
}
\arguments{
\item{plot}{A plot made with ggplot.}
}
\description{
You often want this to share a legend between multiple plots that do not also share axes (i.e. you can't use faceting).
}
\details{
There is (probably much better) functionality in the package ggpubr to use common legend for multiple plots. However, ggpubr depends on latest R version so it can give trouble.
The code was taken from https://stackoverflow.com/questions/13649473/add-a-common-legend-for-combined-ggplots which has a ref to
https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
}
\examples{
## This example is not done yet. But it shows the idea
\dontrun{
mylegend <- extractLegend(p1)
p3 <- grid.arrange(arrangeGrob(p1 + theme(legend.position="none"),
                               p2 + theme(legend.position="none"),
                               nrow=1),
                               mylegend, nrow=2,heights=c(10, 1))
}
}
\seealso{
Other Plotting: \code{\link{NMplotBOV}},
  \code{\link{NMplotBSV}}, \code{\link{NMplotGOF}},
  \code{\link{canvasSize}}, \code{\link{ggIndProfs}},
  \code{\link{ggWater}}, \code{\link{ggstamp}},
  \code{\link{ggwrite}}, \code{\link{theme_pp_facet}},
  \code{\link{theme_pp}}
}
\concept{Plotting}
