% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCovs_df.R
\name{findCovs_df}
\alias{findCovs_df}
\title{Extract columns that do not vary within variables in a data.frame}
\usage{
findCovs_df(data, cols.id = NULL, cols.drop = NULL, debug = F)
}
\arguments{
\item{data}{data.frame in which to look for covariates}

\item{cols.id}{covariates will be searched for in combinations of values in
these columns. Often cols.id will be either empty or ID. But it
can also be both say c("ID","DRUG") or c("ID","TRT").}

\item{cols.drop}{Discard these columns if present.}

\item{debug}{start by running browser()?}
}
\description{
Extract columns that do not vary within variables in a data.frame
}
\examples{
\dontrun{
dat <- NMreadTab("TABLE_OUTPUT.txt")
dat <- as.data.frame(dat)
### very common use
findCovs_(dat,cols.id="ID",cols.drop=c("IRES","TABLE"))
###  an ID column is not needed.
findCovs_df(dat,cols.id=c(),cols.drop=c("IRES","TABLE"))
### need a new data.frame to test for length(cols.id)>1
}
}
\seealso{
Other DataWrangling: \code{\link{NMcompAllRes}},
  \code{\link{NMscanDataDF}}, \code{\link{compareNames}},
  \code{\link{factor2}}, \code{\link{mergeCheck}},
  \code{\link{newNames}}, \code{\link{rbindIntersect}},
  \code{\link{rbindUnion}}, \code{\link{roundDF}},
  \code{\link{roundResp0}}, \code{\link{selectVars}},
  \code{\link{signif2}}, \code{\link{stampObj}},
  \code{\link{tmpcol}}
}
\concept{DataWrangling}
