% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDollar.R
\name{findDollar}
\alias{findDollar}
\title{Find dollar in lines}
\usage{
findDollar(lines, name, keepEmpty = FALSE, keepDollarEmpty = TRUE,
  drop.comments = T)
}
\arguments{
\item{lines}{A Nonmem output or input file as a vector of lines}

\item{name}{The desired dollar name. eg PROBLEM (PROB) without the $}

\item{keepEmpty}{Boolean. Should empty lines be included? Defaults to FALSE.}

\item{keepDollarEmpty}{Boolean. Should a line including the dollar, but with nothing else be kept? Defaults to TRUE.}

\item{drop.comments}{drop comment lines from output.}
}
\value{
The value of desired $name with comments and everything.
}
\description{
Used to find Nonmem parameters in text. This is a background function used by others.
}
\examples{
testLines = list("$asdf  ","$asdf asd ", "asdfa","$bla", "$derp","")
findDollar(testLines,"bla")
findDollar(testLines,"asdf")
findDollar(testLines,"asdf", FALSE, FALSE)
findDollar(testLines,"derp",TRUE)
}
\author{
CIEM
}
\keyword{internal}
