% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCovFun.R
\name{genCovFun}
\alias{genCovFun}
\title{create R function that generates param values based on covariates}
\usage{
genCovFun(NMcode0, pars0, covs0, theta0, omega0, name.fun, debug = F)
}
\description{
create R function that generates param values based on covariates
}
\details{
The function returns a function in plain text. If you want
    to store the code, paste this text to a file. If you want to
    use it right away, do
fun1text <- genCovFun()
newfun <- eval(parse(text=fun1text))
}
