% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIndProfOld.R
\name{ggIndProfOld}
\alias{ggIndProfOld}
\title{Plot individual profiles, split in pages and panels}
\usage{
ggIndProfOld(data, run = "", x = "TIME", dv = "DV", pred = "PRED",
  ipred = "IPRE", grp = "GRP", amt = "AMT", id = "ID",
  xlab = NULL, ylab = NULL, ylab2 = NULL, free.lims = F,
  logy = F, NPerSheet = 12, LLOQ = NULL, use.evid2, facet = id,
  par.prof = NULL, debug = F)
}
\arguments{
\item{run}{The main title of the plot. Called run becaus you often want a Nonmem run name here.}

\item{id}{The name of the subject ID column}

\item{use.evid2}{Should EVID 2 records be used for pred and ipred plotting? The default is to use EVID==2 records if data contains a column called EVID, and this column has at least one value equalling 2.}

\item{facet}{splits plots in pages}

\item{par.prof}{Distinguish multiple profiles in dataset.}
}
\description{
Plot individual profiles, split in pages and panels
}
