% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIndProfs.R
\name{ggIndProfs}
\alias{ggIndProfs}
\title{Plot individual profiles and doses based on NM style dataset}
\usage{
ggIndProfs(data, run, x = "TIME", dv = "DV", pred = "PRED",
  ipred = c("IPRED", "IPRE"), grp, amt = "AMT", id = "ID",
  xlab = NULL, ylab = NULL, ylab2 = NULL, scales = "fixed",
  logy = F, NPerSheet = 12, LLOQ = NULL, use.evid2, facet = id,
  par.prof = NULL, x.inc, grp.label = grp, debug = F)
}
\arguments{
\item{run}{The main title of the plot. Called run becaus you often
want a Nonmem run name here.}

\item{id}{The name of the subject ID column}

\item{use.evid2}{Should EVID 2 records be used for pred and ipred
plotting? The default is to use EVID==2 records if data
contains a column called EVID, and this column has at least
one value equalling 2.}

\item{facet}{splits plots in pages}

\item{par.prof}{Distinguish multiple profiles in dataset.}

\item{grp.label}{Column to use for labeling the sheets (while
sorting by grp). A typical example is that grp is numeric (say
dose including 80 and 280) while grp.label is a character
(including 80 mg and 280 mg). In order to sort correctly, you
must use the numeric variable for grp. But in order to get
nice labels, use the character variable for labels.}
}
\description{
Plot individual profiles and doses based on NM style dataset
}
\details{
The resulting plot object has been limited on x axis by
    coord_cartesian. So if you want to adjust x limits on the
    output from this function, you must use coord_cartesian. xlim
    does not work.
}
