% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggWater.R
\name{ggWater}
\alias{ggWater}
\title{Add watermarks to ggplots}
\usage{
ggWater(text = "Not validated", scale = 1, rot = 30, col = "grey", alpha = 0.5)
}
\arguments{
\item{text}{What should be written on the plot}

\item{scale}{the size}

\item{rot}{rotation of the mark. Don't remember the unit.}
}
\description{
Add watermarks to ggplots
}
\details{
taken from https://www.r-bloggers.com/adding-watermarks-to-plots/
}
\examples{
library(ggplot2)
ff <- qplot(1:10, 11:20) + pmxtricks:::ggWater()
}
\seealso{
Other Plotting: 
\code{\link{NMplotBOV}()},
\code{\link{NMplotBSV}()},
\code{\link{NMplotGOF}()},
\code{\link{canvasSize}()},
\code{\link{extractLegend}()},
\code{\link{ggIndProfs}()},
\code{\link{ggstamp}()},
\code{\link{ggwrite}()},
\code{\link{theme_pp_facet}()},
\code{\link{theme_pp}()}
}
\concept{Plotting}
