% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{means}
\alias{means}
\title{calculate arithmetic or geometric mean and confidence intervals}
\usage{
means(x, type = "arithmetic", na.rm = FALSE, z.rm = FALSE,
  ci = FALSE, dist.ci = "t", p.ci = 0.95, colnames = c("est", "ll",
  "ul"), format = "df", debug = F)
}
\arguments{
\item{x}{vector to calculate the geometric mean of}

\item{type}{type of mean. Default is arithmetic, geometric is available as well. Only first letters needed, so say "geo" or even "g" is enough.}

\item{na.rm}{Remove NA's before doing calculations?}

\item{z.rm}{removes zeros before calculation? Default is FALSE. Can only be TRUE if type="geometric".}

\item{ci}{if TRUE, a data.frame including point estimate and confidence interval returned. If FALSE, a numeric representing the mean value returned.}

\item{dist.ci}{The distribution to use for the confidence interval. Default and only supported is "t". If type=geometric, this is applied after transformation to gaussian.}

\item{p.ci}{probability covered by confidence interval. Default is 0.95}

\item{colnames}{If ci, this defines the column names of the resulting data frame. Default is c("est","ll","ul").}

\item{format}{The format of the result. Possible values are df and num.}

\item{debug}{Start by calling debug()?}
}
\description{
calculate arithmetic or geometric mean and confidence intervals
}
\seealso{
Other Calc: \code{\link{CV}}, \code{\link{geomean}},
  \code{\link{thalf_1a1c}}, \code{\link{thalf_1a2c}},
  \code{\link{thalf_trans_1c}}, \code{\link{trapez}}
}
\concept{Calc}
