% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaCheck.R
\name{metaCheck}
\alias{metaCheck}
\title{Check consistency of a meta data object.}
\usage{
metaCheck(data, meta.data, file.data, match.data = TRUE,
  skip.chars = T, silent = F, debug = F)
}
\arguments{
\item{data}{The dataset being described (i.e. a data.frame).}

\item{meta.data}{The existing meta.data object (see
\code{metaInit}).}

\item{file.data}{Path to the data file.}

\item{match.data}{Check the metadata against the dataset? This is
default and normally a main reason to run this function.}

\item{skip.chars}{Only look at numeric variables in data? Default
is TRUE.}

\item{silent}{Print to terminal whether checks passed. If FALSE,
this info will still be in the returned value.}

\item{debug}{Start by calling browser().}
}
\value{
A logical variable whether all checks passed.
}
\description{
Check consistency of a meta data object.
}
\seealso{
Other DataGen: \code{\link{metaInit}},
  \code{\link{metaPrint}}
}
\concept{DataGen}
