% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaInit.R
\name{metaInit}
\alias{metaInit}
\title{Initialise a metadata object based}
\usage{
metaInit(
  data,
  stamp,
  analysis = NULL,
  header = NULL,
  use.standard.cols = T,
  time.unit = "h",
  dv.unit = NULL,
  amt.unit = NULL,
  ndos.unit = NULL,
  debug = F
)
}
\arguments{
\item{data}{The dataset being described (i.e. a data.frame).}

\item{stamp}{The path to the script that is creating the data.}

\item{analysis}{The name of the analysis (character). This could be a
trial name or any another character string.}

\item{header}{Append lines to the header of the meta data.}

\item{use.standard.cols}{The name is slightly misleading. If FALSE,
only the header will be created. That means that unit arguments
will have no effect.}

\item{time.unit}{The unit of time columns in data. Default is "h"
for hours.}

\item{dv.unit}{If supplied, the unit for DV will be taken from
here. a column called dv.unit will be ignored.}

\item{amt.unit}{If supplied, the unit for AMT will be taken from
here. a column called amt.unit will be ignored.}

\item{ndos.unit}{If supplied, the unit for NDOS will be taken from
here.}

\item{debug}{Start by calling browser()?}
}
\value{
Created meta data object
}
\description{
Create a metadata object based on a dataset and some initial
information. A metadata object is a list containing information
about variables in a dataset plus some more information such as
what dataset is being described, what study it is part of etc.
}
\seealso{
Other DataGen: 
\code{\link{metaCheck}()},
\code{\link{metaPrint}()}
}
\concept{DataGen}
