% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaPrint.R
\name{metaPrint}
\alias{metaPrint}
\title{Print a meta.data object}
\usage{
metaPrint(data, meta.data, file = NULL, file.data, match.data = T,
  silent = !is.null(file), test = F, debug = F)
}
\arguments{
\item{data}{The dataset being described (i.e. a data.frame).}

\item{meta.data}{The existing meta.data object (see \code{metaInit}).}

\item{file}{Path to text file to write to. If not provided,
meta.data is printed to screen.}

\item{file.data}{Path to the data file.}

\item{match.data}{Reorder meta.data to match data? If this function call
is used for saving metadata right after saving the data file, this option
should be used to make sure that the metadata file will match the dataset
as well as possible.}

\item{silent}{If silent, the metadata is not printed to the screen.}

\item{test}{Nice for seeing what the meta data will look like. Problems
will yield warnings, not errors.}

\item{debug}{Start by calling browser()?}
}
\value{
The meta data object printed as formatted text
}
\description{
Print a meta.data object
}
\seealso{
Other DataGen: \code{\link{metaCheck}},
  \code{\link{metaInit}}
}
\concept{DataGen}
