% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOverview.R
\name{modelOverview}
\alias{modelOverview}
\title{goodness of fit and more}
\usage{
modelOverview(path.run, str.output = "NMfolders", folder.out, col.id,
  covs.num, covs.char, var.occ = "OCC", var.occ.char = var.occ,
  covs.num.iov, covs.char.iov, write.output, GRP = "GRP", grp = GRP,
  stamp = NULL, logy = T, LLOQ, args.ggIndProf = NULL,
  grp.trick = NULL, regex.iiv = "^ETAIIV", debug = F)
}
\description{
This function is far from done. The idea is it makes a whole bunch
of plots and tables as an overview of the performance of a nonmem
model.
}
