% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOverviewAll.R
\name{modelOverviewAll}
\alias{modelOverviewAll}
\title{run modelOverview on many runs. But only if it hasn't been done already.}
\usage{
modelOverviewAll(path, debug = F, pattern.dirs = "^[0-9].*",
  force = FALSE, ...)
}
\arguments{
\item{dirs}{Regular expression. Only directories whose name match
this will be taken into account.}
}
\description{
run modelOverview on many runs. But only if it hasn't been done already.
}
