% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newnames.R
\name{newNames}
\alias{newNames}
\title{rename list elements with standard evaluation}
\usage{
newNames(data, names, overwrite = F, skipMissing = TRUE, debug = F)
}
\arguments{
\item{data}{A list (which can be a data.frame)}

\item{names}{A list that contains elements "old" and "new" of
equal length.}

\item{skipMissing}{Just skip if names refers to non-existing elements?}

\item{debug}{Start by calling browser()?}

\item{If}{names$new contains a name of an existing element in
data, should it be overwritten?}
}
\description{
rename list elements with standard evaluation
}
