% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMaddColumns.R
\name{NMaddColumns}
\alias{NMaddColumns}
\title{Merge columns onto all elements in a NM data object}
\usage{
NMaddColumns(data, df, by, debug = F)
}
\arguments{
\item{data}{A NM data object}

\item{df}{a data.frame to merge onto elements of data object.}
}
\description{
Merge columns onto all elements in a NM data object
}
\examples{
df.races <- data.frame(RACE=c(1,3.1),
                       race1=c("White","Japanese"),
                       stringsAsFactors=F)
}
