% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMcheckData.R
\name{NMcheckData}
\alias{NMcheckData}
\title{Check data for Nonmem compatibility}
\usage{
NMcheckData(data, col.id = "ID", col.time = "TIME", col.flagn = NULL, col.row)
}
\arguments{
\item{data}{The data to check}
}
\description{
Check data for Nonmem compatibility
}
