% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMcompAllRes.R
\name{NMcompAllRes}
\alias{NMcompAllRes}
\title{make sure all Nonmem results are saved in rds files}
\usage{
NMcompAllRes(dir, ..., debug = F)
}
\arguments{
\item{dir}{The directory in which to look for nonmem results}

\item{...}{passed to NMcompRes which will pass them to NMscanData}

\item{debug}{start by calling browser()?}
}
\description{
make sure all Nonmem results are saved in rds files
}
