% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMcompAllRes.R
\name{NMcompRes}
\alias{NMcompRes}
\title{Read and store nonmem results in rds}
\usage{
NMcompRes(file, return = TRUE, ...)
}
\arguments{
\item{file}{a nonmem control stream or output (normally, .mod or .lst).}

\item{return}{If TRUE, resulting data structure is invisibly returned. You may want to disable this if the rds is the goal itself. Disabling can save reading the rds if there is nothing to do.}

\item{...}{passed to NMcompRes which will pass them to NMscanData}
}
\description{
Tries to find out if the output rds should be updated and does so if necessary.
}
