% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMextractText.R
\name{NMextractText}
\alias{NMextractText}
\title{NMgetSection generalized with support for lst results}
\usage{
NMextractText(file, lines, text, section, end, char.section,
  char.end = char.section, return = "text", keepEmpty = FALSE,
  keepName = TRUE, keepComments = TRUE, asOne = TRUE,
  simplify = TRUE, cleanSpaces = FALSE, type = "mod", debug = F)
}
\arguments{
\item{file}{A file to read from. Normally a .mod or .lst. See lines also.}

\item{lines}{Text lines to process. This is an alternative to use the fiole
argument.}

\item{section}{The name of section to extract. Examples: "INPUT", "PK",
"TABLE", etc.}

\item{char.end}{A regular expression to capture the end of the section. The
default is to look for the next occurrence of char.section.}

\item{return}{If "text", plain text lines are returned. If "idx", matching
line numbers are returned. "text" is default.}

\item{keepEmpty}{Keep empty lines in output? Default is FALSE.}

\item{keepName}{Keep the section name in output (say, "$PROBLEM") Default is
TRUE. It can only be FALSE, if return"idx".}

\item{keepComments}{Keep comment lines?}

\item{asOne}{If multiple hits, concatenate into one. This will most often be
relevant with name="TABLE". If FALSE, a list will be returned, each
element representing a table. Default is TRUE. So if you want to process
the tables separately, you probably want FALSE here.}

\item{simplify}{If asOne=FALSE, do you want the result to be simplified if
only one table is found? Default is TRUE which is desirable for
interactive analysis. For programming, you probably want FALSE.}

\item{cleanSpaces}{If TRUE, leading and trailing are removed, and multiplied
succeeding white spaces are reduced to single white spaces.}

\item{type}{Either mod, res or NULL. mod is for information that is given in
.mod (.lst can be used but results section is disregarded. If NULL, NA
or empty string, everything is considered.}

\item{debug}{Start by calling browser()?}

\item{char.section.}{The section denoter as a string compatible with regular
expressions. "\\$" for sections in .mod files, "0" for results in .lst
files.}
}
\description{
NMgetSection generalized with support for lst results
}
\details{
This function is planned to get a more general name and then be
    called by NMgetSection.
}
\examples{
NMgetSection(pmxtricks_filepath("examples/nonmem/run001.lst"),section="DATA")


}
\seealso{
Other Nonmem: \code{\link{NMgetSection}},
  \code{\link{NMreadCsv}}, \code{\link{NMreadRun}},
  \code{\link{NMreadTabDF}}, \code{\link{NMreadTab}},
  \code{\link{NMreplacePart}}, \code{\link{NMrunLog}},
  \code{\link{NMscanTables}}, \code{\link{NMtransInput}},
  \code{\link{NMundoReplace}}, \code{\link{NMwriteData}},
  \code{\link{extFormat}}, \code{\link{sumofork}}
}
\concept{Nonmem}
