% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMgetSection.R
\name{NMgetSection}
\alias{NMgetSection}
\title{extract sections of Nonmem control streams}
\usage{
NMgetSection(file, lines, section, return = "text", keepEmpty = FALSE,
  keepName = TRUE, keepComments = TRUE, asOne = TRUE,
  simplify = TRUE, cleanSpaces = FALSE)
}
\arguments{
\item{file}{A file to read from. Normally a .mod or .lst. See lines also.}

\item{lines}{Text lines to process. This is an alternative to use the file
argument.}

\item{section}{The name of section to extract. Examples: "INPUT", "PK",
"TABLE", etc.}

\item{return}{If "text", plain text lines are returned. If "idx", matching
line numbers are returned. "text" is default.}

\item{keepEmpty}{Keep empty lines in output? Default is FALSE.}

\item{keepName}{Keep the section name in output (say, "$PROBLEM") Default is
TRUE. It can only be FALSE, if return"idx".}

\item{keepComments}{Keep comment lines?}

\item{asOne}{If multiple hits, concatenate into one. This will most often be
relevant with name="TABLE". If FALSE, a list will be returned, each
element representing a table. Default is TRUE. So if you want to process
the tables separately, you probably want FALSE here.}

\item{simplify}{If asOne=FALSE, do you want the result to be simplified if
only one table is found? Default is TRUE which is desirable for
interactive analysis. For programming, you probably want FALSE.}

\item{cleanSpaces}{If TRUE, leading and trailing are removed, and multiplied
succeeding white spaces are reduced to single white spaces.}
}
\description{
extract sections of Nonmem control streams
}
