% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMgetSectionOld.R
\name{NMgetSectionOld}
\alias{NMgetSectionOld}
\title{Extract a specific section from Nonmem code}
\usage{
NMgetSectionOld(file, lines, name, keepEmpty = FALSE,
  keepDollarEmpty = TRUE, dropComments = T, return = "text")
}
\arguments{
\item{lines}{A Nonmem output or input file as a vector of lines}

\item{name}{The desired section name. eg PROBLEM (PROB) without the
$}

\item{keepEmpty}{Boolean. Should empty lines be included? Defaults
to FALSE.}

\item{keepDollarEmpty}{Boolean. Should a line including the dollar,
but with nothing else be kept? Defaults to TRUE.}

\item{dropComments}{drop comment lines from output.}

\item{return}{"text" or "lineNumbers". What do you want to get back
from the function? The actual text or the line numbers to find
the text?}
}
\value{
The value of desired $name with comments and everything.
}
\description{
Extract a specific section from Nonmem code
}
\examples{
testLines = list("$asdf  ","$asdf asd ", "asdfa","$bla", "$derp","")
NMgetSection(testLines,"bla")
NMgetSection(testLines,"asdf")
NMgetSection(testLines,"asdf", FALSE, FALSE)
NMgetSection(testLines,"derp",TRUE)
}
