% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMplotBOV.R
\name{NMplotBOV}
\alias{NMplotBOV}
\title{Generate distribution plots of between-occasion variability terms from
Nonmem}
\usage{
NMplotBOV(data, regex.eta = "^ETABOV", col.id = "ID", covs.num.iov,
  covs.char.iov, var.occ = "OCC", var.occ.char = var.occ,
  fun.file = identity, save = F, stamp = NULL, debug = F)
}
\arguments{
\item{data}{A dataset - will be converted to data.frame so data.table is OK.}

\item{regex.eta}{A regular expression defining the naming of the ETA's of
interest.}

\item{col.id}{The name of the id column in data. Default is ID like Nonmem.}

\item{covs.num.iov}{Names of columns containing numerical covariates to plot
the random effects against.}

\item{covs.char.iov}{Names of columns containing categorical covariates to
plot the random effects against.}

\item{var.occ}{The name of the numerical occasion variable (if the model has
say OCC=1:10).}

\item{var.occ.char}{The name of the character representation of occasions,
if one exists. This column could contain "Screening", "Visit 2",
"Visit 3" etc.}

\item{fun.file}{If saving plots, this function can be used to translate the
file names. The inputs given to the function argument are
"iov_pairs.png" and "iov_covs_n.png".}

\item{save}{Save the generated plots?}

\item{stamp}{If saving the plots, a stamp to add. See ggstamp.}

\item{debug}{Start by running browser()?}
}
\description{
Generate distribution plots of between-occasion variability terms from
Nonmem
}
\seealso{
NMplotBSV
}
