% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMplotGOF.R
\name{NMplotGOF}
\alias{NMplotGOF}
\title{A standard panel of residual plots}
\usage{
NMplotGOF(
  data,
  res = CWRES,
  ipre = IPRED,
  time = TIME,
  colour = NULL,
  arrange = T,
  title,
  debug = F
)
}
\arguments{
\item{data}{is all rows from the Nonmem output. Only the EVID==0
subset will be used.}

\item{res}{The individual residuals (how to derive them from data)}

\item{ipre}{expression for individual predictions}

\item{time}{Time variable (normally TIME or NOMTIME or TAPD)}

\item{colour}{expression for colour mapping}

\item{arrange}{If true, the plots will be arranged with
arrangeGrob. If not, they will be returned as individual plots
in a list.}

\item{title}{An optional character title for the plots.}

\item{debug}{Start by runing bowser?}
}
\description{
A standard panel of residual plots
}
\details{
All parameters must be given as expressions (no quotes)
}
\seealso{
Other Plotting: 
\code{\link{NMplotBSV}()},
\code{\link{canvasSize}()},
\code{\link{extractLegend}()},
\code{\link{ggIndProfs}()},
\code{\link{ggWater}()},
\code{\link{ggstamp}()},
\code{\link{ggwrite}()},
\code{\link{theme_pp_facet}()},
\code{\link{theme_pp}()}
}
\concept{Plotting}
