% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadRun.R
\name{NMreadRun}
\alias{NMreadRun}
\title{read model details}
\usage{
NMreadRun(run, debug = F)
}
\arguments{
\item{run}{path to an lst file.}

\item{debug}{Start by calling browser()?}
}
\description{
This function is used to produce run logs or run records by NMrunLog. Much better alternatives can most likely be found elsewhere.
}
\seealso{
Other Nonmem: \code{\link{NMgetSection}},
  \code{\link{NMreadCsv}}, \code{\link{NMreadTab}},
  \code{\link{NMreplacePart}}, \code{\link{NMrunLog}},
  \code{\link{NMscanTables}}, \code{\link{NMtransInput}},
  \code{\link{NMundoReplace}}, \code{\link{NMwriteData}},
  \code{\link{extFormat}}, \code{\link{sumofork}}
}
\concept{Nonmem}
