% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadTab.R
\name{NMreadTab}
\alias{NMreadTab}
\title{Read a table file from NONMEM}
\usage{
NMreadTab(file, silent = F, skip, sep = " ", debug = F)
}
\arguments{
\item{file}{path to NONMEM table file}

\item{silent}{logical stating whether or not information is printed.}

\item{skip}{Number of lines skipes in the start of the file. Some files have
additional headers so if table looks wierd try adding 1 to skip. Notice
that the lines TABLE 1 etc from Nonmem will be handled automatically.}

\item{sep}{The seperator to use in read.table. If " " is used, any number of blanks will work as one separator in the datafile. This should be OK for Nonmem output as the function is intended for but may be a problem for other tables.}
}
\value{
Nonmem table as df.
}
\description{
Read a table file from NONMEM
}
\author{
Philip Delff
}
