% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreplacePart.R
\name{NMreplacePart}
\alias{NMreplacePart}
\title{replace ("dollar") parts of a nonmem control stream}
\usage{
NMreplacePart(path, dollar, newlines, newpath, backup = T,
  blank.append = T, test = F, debug = F)
}
\arguments{
\item{path}{The run to edit. If a directory is given, the file is assumed to
be called input.txt in that folder.}

\item{dollar}{The name of the dollar section to update. Example:
dollar="EST". See findDollar.}

\item{newlines}{The new text. Better be broken into lines in a character
vector since this is simply written with writeLines.}

\item{newpath}{path to new run. If missing, path is used. If NULL, output is
returned rather than written.}

\item{backup}{In case you are overwriting the old file, do you want to
backup the file (to say, backup_input.txt)?}

\item{test}{Want to see the resulting input.txt and not write it to disk? Default is FALSE.}

\item{debug}{start by running browser()?}
}
\description{
Just give the dollar name, the new lines and the file path, and the
"dollar section", and the input to Nonmem will be updated.
}
\details{
The new file will be written with unix-style line endings.
}
\examples{
newlines <- "$EST POSTHOC INTERACTION METHOD=1 NOABORT PRINT=5 MAXEVAL=9999 SIG=3"
nmReplacePart(path="run01.mod",dollar="EST", newlines=newlines)
}
