% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMrunLog.R
\name{NMrunLog}
\alias{NMrunLog}
\title{get an overview of your Nonmem runs}
\usage{
NMrunLog(dir, runs, runs.omit, debug = F)
}
\arguments{
\item{dir}{The directory in which to search for runs}

\item{runs}{Runs to include in case you only want a specific subset.}

\item{runs.omit}{Runs to omit in case you want all but these.}
}
\description{
get an overview of your Nonmem runs
}
