% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMscan.R
\name{NMscan}
\alias{NMscan}
\title{automatically find Nonmem tables and organize data}
\usage{
NMscan(path.run, col.id = "ID", col.row = "ROW", col.grp = NULL,
  col.occ = "OCC", structure = "full", use.input = T,
  reconstructRows = F, debug = F)
}
\arguments{
\item{col.grp}{If present, ID and OCC level info is grouped by col.grp. So should only be needed for cross-over.}
}
\description{
automatically find Nonmem tables and organize data
}
