% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMscanTables.R
\name{NMscanTables}
\alias{NMscanTables}
\title{read all output data tables in nonmem run}
\usage{
NMscanTables(file, details = F)
}
\arguments{
\item{file}{the nonmem file to read (normally .mod or .lst)}

\item{details}{If TRUE, metadata is added to output. In this case, you get a list. I'd say, enable if doing programming.}
}
\value{
A list of all the tables as data.frames. If details=TRUE, this is in one element, called data, and meta is another element. If not, only the element corresponding to data is returned.
}
\description{
read all output data tables in nonmem run
}
