% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMwriteData.R
\name{NMwriteData}
\alias{NMwriteData}
\title{Write dataset for use in Nonmem (and R)}
\usage{
NMwriteData(data, file, drop = NULL, drop.lowercase = FALSE,
  write.csv = TRUE, write.RData = F, write.rds = write.csv,
  force.row = FALSE, script, debug = FALSE)
}
\arguments{
\item{data}{The dataset to write to Nonmem.}

\item{file}{The file to write to.}

\item{drop}{Specific columns to drop before writing the data to file.}

\item{drop.lowercase}{Drop all columns which names containe lowercase letters}

\item{write.csv}{Write to csv file?}

\item{write.RData}{In case you want to save to .RData object. Not recommended. Use write.rds instead.}

\item{write.rds}{write an rds file?}

\item{force.row}{Ensure that data contains a ROW column counting the rows in the dataset, and add one if none exists. Defaults to FALSE (default is not editing the dataset at all).}

\item{script}{If provided, the object will be stamped with this script name before saved to rds.}

\item{debug}{Start by running browser()?}
}
\description{
Instead of trying to remember the arguments to pass to write.csv, use this
wrapper. It does much more. It tells you what to write in $DATA and $INPUT
in nonmem, and it exports a rds file as well which is highly preferable for
use in R.
}
\seealso{
Other Nonmem: \code{\link{NMaddColumns}},
  \code{\link{NMtransInput}}, \code{\link{NMundoReplace}}
}
\concept{Nonmem}
