% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objInfo.R
\name{objInfo}
\alias{objInfo}
\title{Get stamped info about object}
\usage{
objInfo(data)
}
\arguments{
\item{data}{The dataset to receive info about}
}
\description{
Get stamped info about object
}
\seealso{
stampObj

Other DataWrangling: \code{\link{NMaddColumns}},
  \code{\link{NMcompAllRes}}, \code{\link{NMorderColumns}},
  \code{\link{NMscanData}}, \code{\link{compareNames}},
  \code{\link{factor2}}, \code{\link{findCovs}},
  \code{\link{mergeCheck}}, \code{\link{newNames}},
  \code{\link{rbindIntersect}}, \code{\link{rbindUnion}},
  \code{\link{roundDF}}, \code{\link{roundResp0}},
  \code{\link{selectVars}}, \code{\link{signif2}},
  \code{\link{stampObj}}, \code{\link{tmpcol}}
}
\concept{DataWrangling}
