% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMorderColumns.R
\name{orderColumns}
\alias{orderColumns}
\title{Order columns in dataset for use in Nonmem.}
\usage{
orderColumns(data, first, last, debug = F)
}
\arguments{
\item{first}{Columns that should come before alphabetic sorting. Default is c("ROW","ID","NTIM","TIME","EVID","CMT","AMT","RATE","DV","MDV","FLAG","OCC","ROUTE","GRP","TRIAL")}

\item{last}{Columns to sort after alphabetic ordering. Default is none.}

\item{debug}{Start by calling browser()?}
}
\description{
Order the columns in a data.frame for easy export to Nonmem. Standard
columns like "ROW", "ID", "NTIM", etc. will be first by default, then other
capital case named columns, then lowercase named columns (one or more
lowercase letter means that column is sorted as lowercase). Except for
columns mentioned in "first" and "last" arguments, columns are sorted
alphabetically (after by case). In sohort, priority is 1: Case, first/last,
alphabetical. This means that last="BW" will put body weight as last of
capital, but before lowercase columns.
}
\author{
Philip Delff
}
