% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpdSimMany.R
\name{pkpdSimMany}
\alias{pkpdSimMany}
\title{Simulate multiple subjects with PKPDsim}
\usage{
pkpdSimMany(model, doses, pars, simtimes, col.id, debug = F, ...)
}
\arguments{
\item{doses}{amt is in mg.}
}
\description{
Simulate multiple subjects with PKPDsim
}
