% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{pksim1}
\alias{pksim1}
\title{A very simple simulated PK dataset}
\format{
A data.table with 76 rows, 6 columns. 4 subjects, 19 observations each.
\describe{
\item{ROW}{A row identifier}
\item{ID}{Subject ID.}
\item{TIME}{Time since first dose.}
\item{EVID}{0 for dose.}
\item{CMT}{Compartment identifier. 2 for central compartment.}
\item{DV}{Observation value.}
\item{MDV}{Is observation missing?}
}
}
\source{
simulated
}
\usage{
pksim1
}
\description{
This dataset is mainly used to illustrate functionality of functions in the package.
}
\keyword{datasets}
