% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxtricks_utils.R
\name{pmxtricks_filepath}
\alias{pmxtricks_filepath}
\title{get a file or dir from a the pmxtricks package}
\usage{
pmxtricks_filepath(...)
}
\arguments{
\item{...}{parameters to pass to system.file}
}
\description{
get a file or dir from a the pmxtricks package
}
\details{
a light wrapper around system.file
}
