% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantbin.R
\name{quantbin}
\alias{quantbin}
\title{Bin observations by quantiles. Label by bin number or by interval.}
\usage{
quantbin(x, nbins, label = "num", na.rm = FALSE, ...)
}
\arguments{
\item{x}{The observations}

\item{nbins}{Number of bins to use}

\item{label}{label="num" gives a numeric bin number
(findInterval). label="interval" gives a character
representation of the interval (cut).}

\item{na.rm}{na.rm=T is needed for quantile to be able to estimate
the distribution if x contains NA's. Notice, if na.rm=T, an NA
element in x will still result in an NA.}

\item{...}{additional arguments passed to quantile.}

\item{type}{passed to quantile}
}
\description{
This is simple stuff, but I can never remember the exact quantile
and findInterval/cut commands to use. quantbin finds quantiles
using quantile and then assigns bins using either findInterval or
cut.
}
