% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindIntersect.R
\name{rbindIntersect}
\alias{rbindIntersect}
\title{rbind data.frames by using only common columns}
\usage{
rbindIntersect(...)
}
\arguments{
\item{...}{data frames to stack. Must be of class data.frame.}
}
\value{
a data.frame containing all data
}
\description{
rbind data.frames by using only common columns
}
\examples{
x <- data.frame(a=1:3)
y <- data.frame(b=1:3,c="h")
z <- data.frame(a=4:5,b=letters[4:5],d=c(NA,5))
rbindIntersect(x,y,z)
}
\seealso{
Other DataWrangling: \code{\link{NMaddColumns}},
  \code{\link{NMcompAllRes}}, \code{\link{NMorderColumns}},
  \code{\link{NMscanDataDF}}, \code{\link{NMscanData}},
  \code{\link{compareNames}}, \code{\link{factor2}},
  \code{\link{findCovs_df}}, \code{\link{findCovs}},
  \code{\link{findVars}}, \code{\link{mergeCheck}},
  \code{\link{newNames}}, \code{\link{objInfo}},
  \code{\link{rbindUnion}}, \code{\link{roundDF}},
  \code{\link{roundResp0}}, \code{\link{selectVars}},
  \code{\link{signif2}}, \code{\link{stampObj}},
  \code{\link{tmpcol}}
}
\concept{DataWrangling}
