% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindUnion.R
\name{rbindUnion}
\alias{rbindUnion}
\title{rbind data.frames by filling in NA's}
\usage{
rbindUnion(...)
}
\arguments{
\item{...}{data frames to stack. Must be of class data.frame.}
}
\value{
a data.frame containing all data
}
\description{
This function is an implementation of the functionality as in rbind.fill in
plyr. The intention of writing this function is to avoid being dependent on
plyr. There is no advantage of using this function instead of
plyr::rbind.fill or dplyr::bind_rows.
}
\examples{
x <- data.frame(a=1:3)
y <- data.frame(b=1:3,c="h")
z <- data.frame(a=4:5,b=letters[4:5],d=c(NA,5))
rbindUnion(x,y,z)
}
\author{
philip@delff.dk
}
