% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPsnVpc.R
\name{readPsnVpc}
\alias{readPsnVpc}
\title{read vpc output from psn}
\usage{
readPsnVpc(vpc.results = NULL, npc.results = NULL, verbose = FALSE, ...)
}
\arguments{
\item{vpc.results}{The name of the results file from running the PsN command vpc. Often this
is named 'vpc\_results.csv'.}

\item{npc.results}{The name of the results file from running the PsN command npc. Often this is
named 'npc\_results.csv'. relative or absolute paths to the file are allowed
as for vpc.results.}

\item{verbose}{Text messages passed to screen or not.}

\item{...}{arguments passed to other functions.}
}
\description{
Taken from Based on function from package:xpose4 (read.npc.vpc.results)
This function reads in a results file from running either the PsN command vpc or npc. The function
then parses the file and passes the result to plotting functions
}
