% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVPC.R
\name{readVPC}
\alias{readVPC}
\title{read VPC results a little more flexibelly than with xpose4::read.npc.vpc.results}
\usage{
readVPC(file, ifMissing = NULL, debug = F)
}
\arguments{
\item{file}{The file to read.}

\item{ifMissing}{The value to return if file is missing.}

\item{debug}{Start by calling debug()?}
}
\description{
read VPC results a little more flexibelly than with xpose4::read.npc.vpc.results
}
\concept{Nonmem.
importFrom xpose4 read.npc.vpc.results}
