% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundDF.R
\name{roundDF}
\alias{roundDF}
\title{Report numbers with a fixed formatting.}
\usage{
roundDF(x, FUN = signif2, ...)
}
\arguments{
\item{x}{a numeric vector or a data.frame with some of the columns numeric.}

\item{FUN}{the rounding function to apply to numeric columns.}

\item{...}{Arguments passed to FUN}
}
\description{
sigdigs is a wrapper that rounds whatever numeric in a data.frame
(because round(df) returns an error if df contains
non-numerics). You often want to run this function before creating
a csv file as output.
}
\examples{
df1 <- data.frame(obs=letters[1:3],x1=rnorm(3),x2=c(1.24/1e4,1.1334e6,1.1),x3=1:3)
roundDF(df1,digits=3)
}
\seealso{
Other DataWrangling: \code{\link{NMcompAllRes}},
  \code{\link{NMscanDataDF}}, \code{\link{compareNames}},
  \code{\link{factor2}}, \code{\link{findCovs_df}},
  \code{\link{mergeCheck}}, \code{\link{newNames}},
  \code{\link{rbindIntersect}}, \code{\link{rbindUnion}},
  \code{\link{roundResp0}}, \code{\link{selectVars}},
  \code{\link{signif2}}, \code{\link{stampObj}},
  \code{\link{tmpcol}}
}
\concept{DataWrangling}
