% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundResp0.R
\name{roundResp0}
\alias{roundResp0}
\title{round to a given number of digits - and respect zeros.}
\usage{
roundResp0(x, digits)
}
\arguments{
\item{x}{A vector to round.}

\item{digits}{Number of digits to}
}
\description{
round to a given number of digits - and respect zeros.
}
\seealso{
Other DataWrangling: 
\code{\link{compareNames}()},
\code{\link{factor2}()},
\code{\link{mergeCheck}()},
\code{\link{newNames}()},
\code{\link{rbindIntersect}()},
\code{\link{rbindUnion}()},
\code{\link{roundDF}()},
\code{\link{selectVars}()},
\code{\link{signif2}()},
\code{\link{stampObj}()}
}
\concept{DataWrangling}
