% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVars.R
\name{selectVars}
\alias{selectVars}
\title{Select only wanted columns - If any of them are missing, tell the user
which ones.}
\usage{
selectVars(data, vars)
}
\arguments{
\item{data}{data.frame or data.table to select columns from.}

\item{vars}{columns to select. A vector of characters.}
}
\description{
This is just a wrapper to data[,vars]. But first it checks if vars are in
data. And if not it is reported in an error to the user. This saves you the
manual investigation every time data[,vars] gives you the annoyingly
non-informative error "undefined columns selected".
}
\seealso{
Other DataWrangling: \code{\link{NMaddColumns}},
  \code{\link{NMcompAllRes}}, \code{\link{NMorderColumns}},
  \code{\link{NMscanData}}, \code{\link{compareNames}},
  \code{\link{factor2}}, \code{\link{findCovs}},
  \code{\link{mergeCheck}}, \code{\link{newNames}},
  \code{\link{objInfo}}, \code{\link{rbindIntersect}},
  \code{\link{rbindUnion}}, \code{\link{roundDF}},
  \code{\link{roundResp0}}, \code{\link{signif2}},
  \code{\link{stampObj}}, \code{\link{tmpcol}}
}
\concept{DataWrangling}
