% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stampObj.R
\name{stampObj}
\alias{stampObj}
\title{stamp a dataset or any other object}
\usage{
stampObj(data, script, ...)
}
\arguments{
\item{data}{The dataset to stamp.}

\item{script}{path to the script where the dataset was generated.}

\item{...}{other named metadata elements to add to the dataset. Example:
Description="PK data for ph1 trials in project".}
}
\description{
Dataset metadata makes it possible to trace an archived dataset back to the
code that generated it. The metadata added by stampData can be accessed
using the function dsInfo.
}
\seealso{
objInfo

Other DataWrangling: \code{\link{NMaddColumns}},
  \code{\link{NMcompAllRes}}, \code{\link{NMorderColumns}},
  \code{\link{NMscanData}}, \code{\link{compareNames}},
  \code{\link{factor2}}, \code{\link{findCovs_df}},
  \code{\link{findCovs}}, \code{\link{findVars}},
  \code{\link{mergeCheck}}, \code{\link{newNames}},
  \code{\link{objInfo}}, \code{\link{rbindIntersect}},
  \code{\link{rbindUnion}}, \code{\link{roundDF}},
  \code{\link{roundResp0}}, \code{\link{selectVars}},
  \code{\link{signif2}}, \code{\link{tmpcol}}
}
\concept{DataWrangling}
