% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofork.R
\name{sumofork}
\alias{sumofork}
\title{Platform independent sumo}
\usage{
sumofork(model, tableType = 2, format.estimate = "\% -#6.4g",
  format.rse = "\%#6.3g")
}
\arguments{
\item{model}{A .mod, .lst, .ext or .cov file.}

\item{tableType}{Passed to extTransform}

\item{format.estimate}{How to format the estimated parameter values.}

\item{format.rse}{How to format the relative standard error estimates.}
}
\description{
Platform independent sumo
}
\details{
This is a fork of nonmem2R::sumo with very little
    modification. Many thanks to Magnus for a great job!
}
\seealso{
Other Nonmem: \code{\link{NMextractText}},
  \code{\link{NMgetSection}}, \code{\link{NMreadTabDF}},
  \code{\link{NMreplacePart}}, \code{\link{NMundoReplace}},
  \code{\link{extFormat}}
}
\concept{Nonmem}
