% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_thalf.R
\name{thalf.1a.2c}
\alias{thalf.1a.2c}
\title{1 absorption depot, 2 compartments}
\usage{
thalf.1a.2c(pars, cl = "CL", ka = "KA1", v2 = "V2", q = "Q",
  v3 = "V3", mutate = F, debug = F)
}
\description{
Can take multiple parameter sets and return a data.frame of halftimes.
}
