% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_thalf.R
\name{thalf_1a1c}
\alias{thalf_1a1c}
\title{Thalf for 1 absorption depot, 1 compartment}
\usage{
thalf_1a1c(pars, cl = "CL", ka = "KA1", v2 = "V2", debug = F)
}
\arguments{
\item{pars}{A parameter table. Thalf will be calculated for all rows.}

\item{cl}{The name of the column with clearance values. Can represent CL/F if volume variables do the same.}

\item{ka}{The name of the column representing absorption rate.}

\item{v2}{The name of the column with central volume (see cl as well).}

\item{debug}{Start by calling browser()?}
}
\description{
Can take multiple parameter sets and return a data.frame of halftimes.
}
\seealso{
Other Calc: 
\code{\link{thalf_1a2c}()},
\code{\link{thalf_trans_1c}()}
}
\concept{Calc}
