% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_thalf.R
\name{thalf_1a2c}
\alias{thalf_1a2c}
\title{Thalf for 1 absorption depot, 2 compartments}
\usage{
thalf_1a2c(
  pars,
  cl = "CL",
  ka = "KA1",
  v2 = "V2",
  q = "Q",
  v3 = "V3",
  transform = F,
  debug = F
)
}
\arguments{
\item{pars}{A parameter table. Thalf will be calculated for all rows.}

\item{cl}{The name of the column with clearance values. Can represent CL/F if volume variables do the same.}

\item{ka}{The name of the column representing absorption rate.}

\item{v2}{The name of the column with central volume (see cl as well).}

\item{q}{The name of the column with intercompartmental clearance.}

\item{v3}{The name of the column with peripharel volume (see cl as well).}

\item{transform}{Add results to existing dataset?}

\item{debug}{Start by calling browser()}
}
\description{
Can take multiple parameter sets and return a data.frame of halftimes.
}
\seealso{
Other Calc: 
\code{\link{CV}()},
\code{\link{means}()},
\code{\link{thalf_1a1c}()},
\code{\link{thalf_trans_1c}()},
\code{\link{trapez}()}
}
\concept{Calc}
