% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pp.R
\name{theme_pp}
\alias{theme_pp}
\title{A custom theme - aim is powerpoint}
\usage{
theme_pp(theme0 = ggplot2::theme_bw, ...)
}
\arguments{
\item{theme0}{The theme to modify}

\item{...}{arguments passed to theme() as last thing.}
}
\description{
A custom theme - aim is powerpoint
}
\seealso{
Other plotting: \code{\link{NMplotFit}},
  \code{\link{extractLegend}}, \code{\link{theme_pp_facet}}
}
\concept{plotting}
