% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pp.R
\name{theme_pp_facet}
\alias{theme_pp_facet}
\title{a version of theme_pp suited for facet'ed plots}
\usage{
theme_pp_facet(theme0 = theme_bw, ...)
}
\arguments{
\item{theme0}{The theme to modify}
}
\description{
a version of theme_pp suited for facet'ed plots
}
\seealso{
Other plotting: \code{\link{NMplotFit}},
  \code{\link{extractLegend}}, \code{\link{theme_pp}}
}
\concept{plotting}
