% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpcol.R
\name{tmpcol}
\alias{tmpcol}
\title{generate a name for a new data column that is not already in use.}
\usage{
tmpcol(data, base = "atmpcol999", max.it = 100)
}
\arguments{
\item{data}{The dataset to find a new element name for}

\item{base}{The base name of the new element. A number will appended to this string that will ensure that the new element name is not already in use.}

\item{max.it}{Maximum number of iterations on element name.}
}
\description{
generate a name for a new data column that is not already in use.
}
\seealso{
Other DataWrangling: \code{\link{NMaddColumns}},
  \code{\link{NMcompAllRes}}, \code{\link{NMorderColumns}},
  \code{\link{NMscanDataDT}}, \code{\link{NMscanData}},
  \code{\link{compareNames}}, \code{\link{factor2}},
  \code{\link{findCovs_df}}, \code{\link{findCovs}},
  \code{\link{findVars}}, \code{\link{mergeCheck}},
  \code{\link{newNames}}, \code{\link{objInfo}},
  \code{\link{rbindIntersect}}, \code{\link{rbindUnion}},
  \code{\link{roundDF}}, \code{\link{roundResp0}},
  \code{\link{selectVars}}, \code{\link{signif2}},
  \code{\link{stampObj}}
}
\concept{DataWrangling}
