% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapez.R
\name{trapez}
\alias{trapez}
\title{trapezoidal AUC on linear scale}
\usage{
trapez(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{The vector to integrate y with respect to (typically TIME to get AUC).}

\item{y}{The variable to integrate.}

\item{na.rm}{Remove indexes in x and y wherever x or y are NA.}
}
\description{
This is a numerical integration of y with respect to x by the trapezoidal method on linear scale.
}
\seealso{
Other Calc: \code{\link{CV}}, \code{\link{geomean}},
  \code{\link{means}}, \code{\link{thalf_1a1c}},
  \code{\link{thalf_1a2c}}, \code{\link{thalf_trans_1c}}
}
\concept{Calc}
