% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{Calculate coefficient of variation of data}
\usage{
CV(x, log = F)
}
\arguments{
\item{x}{The data @param log If TRUE, the geometric coefficient of variation
is calculated. This is sqrt(exp(var(log(x))-1).}

\item{log}{If true, a geometric CV is derived.}
}
\description{
Calculate coefficient of variation of data
}
\details{
This function is intended to be used on data. For a log-normal
    THETA[1]*EXP(OMEGA[1]) Nonmem ## parameter, do
    CV=sqrt(exp(OMEGA[1,1])-1).
importFrom stats var sd
}
\seealso{
Other Calc: \code{\link{geomean}}, \code{\link{means}},
  \code{\link{thalf_1a1c}}, \code{\link{thalf_1a2c}},
  \code{\link{thalf_trans_1c}}, \code{\link{trapez}}
}
\concept{Calc}
