% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extFormat.R
\name{extFormat}
\alias{extFormat}
\title{there seems to be a bug in nonmem2R::extFormat. This function tries to solve that.}
\usage{
extFormat(ext, format.estimate = "\% -#6.4g", format.rse = "\%#6.3g")
}
\arguments{
\item{ext}{the result ofname of the lst file without the .lst extension.}

\item{format.estimate}{format for estimated value, passed to sprintf}

\item{format.rse}{format for rse, passed to sprintf}
}
\description{
The code is copied from nonmem2R and only slightly modified.
}
\seealso{
Other Nonmem: \code{\link{NMextractText}},
  \code{\link{NMgetSection}}, \code{\link{NMreadRun}},
  \code{\link{NMreadTabDF}}, \code{\link{NMreplacePart}},
  \code{\link{NMrunLog}}, \code{\link{NMundoReplace}},
  \code{\link{sumofork}}
}
\concept{Nonmem}
