% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2.R
\name{factor2}
\alias{factor2}
\title{factor with automated levels by sorting a corresponding variable}
\usage{
factor2(x, by = x, debug = F)
}
\arguments{
\item{x}{The variable make a factor variable out of.}

\item{by}{The variable which values the levels of the resulting values will
be sorted by.}

\item{debug}{start by calling browser()?}
}
\description{
factor with automated levels by sorting a corresponding variable
}
\details{
Use base::reorder instead. I think it does the same. factor2 is
    sometimes convenient, but it does not work within data.table.
}
