% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{calculate geometric  mean}
\usage{
geomean(x, z.rm = F)
}
\arguments{
\item{x}{vector to calculate the geometric mean of}

\item{z.rm}{removes zeros before calculation?}
}
\description{
calculate geometric  mean
}
