% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstamp.R
\name{ggstamp}
\alias{ggstamp}
\title{Always stamp your plots with script name}
\usage{
ggstamp(plot, stamp = "no stamp", time = Sys.time())
}
\arguments{
\item{plot}{The plot to be stamped.}

\item{stamp}{the script name. Date and time will be added
automatically.}
}
\value{
the plot with a stamp
}
\description{
This function is used to stamp ggplot type plots with datetime and
script name. User must provide the script name. Notice that ggplot
2.2.1 or newer is required.
}
\examples{
norms <- do.call(rbind,lapply(1:3,function(mu)data.frame(POP=mu,population=paste("Population",mu),parameter=rnorm(seq(mu-6,mu+6,.01),mean=mu,sd=sin(mu)*2))))
## A very quick way to see densities is using ggplot
p1 <- ggplot(norms,aes(parameter,fill=population))+geom_density(alpha=.5)
gsave(p1)
stamp <- "note"
ggstamp(p1,stamp)
## Or use gsave which will call ggstamp automatically.
gsave(p1,stamp=stamp,canvas="wide")
}
