% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggWater.R
\name{ggWater}
\alias{ggWater}
\title{get watermarks in ggplots}
\usage{
ggWater(text = "Not validated", scale = 1, rot = 0)
}
\arguments{
\item{text}{What should be written on the plot}

\item{scale}{the size}

\item{rot}{rotation of the mark. Don't remember the unit.}
}
\description{
get watermarks in ggplots
}
\details{
taken from https://www.r-bloggers.com/adding-watermarks-to-plots/
}
\examples{
ff <- qplot(1:10, 11:20) + qcp_water()
}
