% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwrite.R
\name{ggwrite}
\alias{ggwrite}
\title{Export plots created with ggplot (and more) to files (png or pdf) - or show them on screen.}
\usage{
ggwrite(plot, file, stamp, canvas = "standard", onefile = F,
  res = 200, debug = F, paper = "special", save = T)
}
\arguments{
\item{plot}{A plot object or a list of plots. Normally generated with ggplot
or qplot. But it can also be from grid.arrange or arrangeGrob with class
gtable. That is experimental though. Not sure exactly what classes are
supported.}

\item{file}{A file to export to. Must end in .png or .pdf. If plot is a
list, see onefile. If missing, plot is shown on screen.}

\item{stamp}{This should normally be the path to your script. Requires
ggplot >=2.2.1.}

\item{canvas}{Can either be a character with one of the values "standard",
"wide", "A4", "xeqy","wide-screen" or it can be a list with elements
width and height with single values (unit is inches).}

\item{onefile}{Only applicable if plot is a list. If plot is a list and
onefile=T, all plots will be put in a pdf (file must end in pdf) with
one plot per page. If plot is a list and onefile=F, numbered files will
be created - one per list element.}

\item{res}{Resolution. Passed to png.}

\item{debug}{If TRUE, browser is called to begin with.}

\item{paper}{Only used with pdf device. See ?pdf.}

\item{save}{Save the plot to the given file or just show? Defaults to
TRUE. Hint, if you use an "exportFlag", use save=exportFlag.}
}
\value{
Nothing.
}
\description{
Export plots created with ggplot (and more) to files (png or pdf) - or show them on screen.
}
\examples{
exportFlag <- T
norms <- do.call(rbind,lapply(1:3,function(mu)data.frame(POP=mu,population=paste("Population",mu),parameter=rnorm(seq(mu-6,mu+6,.01),mean=mu,sd=sin(mu)*2))))
## A very quick way to see densities is using ggplot
p1 <- ggplot(norms,aes(parameter,fill=population))+geom_density(alpha=.5)
ggwrite(p1)  ## view plot on screen
stamp <- "note"
ggwrite(p1,stamp=stamp,canvas="wide",save=exportFlag)
}
