% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMaddColumns.R
\name{NMaddColumns}
\alias{NMaddColumns}
\title{Merge columns onto all elements in a NM data object}
\usage{
NMaddColumns(data, df, by, debug = FALSE)
}
\arguments{
\item{data}{A NM data object}

\item{df}{a data.frame to merge onto elements of data object.}

\item{by}{passed to merge}

\item{debug}{Start by calling browser()?}
}
\description{
This is useful for adding columns to a nonmem data object as
    created by NMscanData because it does so to all elements containing the
    by column in the object. The merge call is run by mergeCheck with all.x=TRUE.
}
\examples{
df.races <- data.frame(RACE=c(1,3.1),
                       race1=c("White","Japanese"),
                       stringsAsFactors=FALSE)
}
\seealso{
Other DataWrangling: \code{\link{NMcompAllRes}},
  \code{\link{NMorderColumns}}, \code{\link{NMscanDataDF}},
  \code{\link{NMscanData}}, \code{\link{compareNames}},
  \code{\link{factor2}}, \code{\link{findCovs_df}},
  \code{\link{findCovs}}, \code{\link{findVars}},
  \code{\link{mergeCheck}}, \code{\link{newNames}},
  \code{\link{objInfo}}, \code{\link{rbindIntersect}},
  \code{\link{rbindUnion}}, \code{\link{roundDF}},
  \code{\link{roundResp0}}, \code{\link{selectVars}},
  \code{\link{signif2}}, \code{\link{stampObj}},
  \code{\link{tmpcol}}
}
\concept{DataWrangling}
