% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMcompAllRes.R
\name{NMcompAllRes}
\alias{NMcompAllRes}
\title{make sure all Nonmem results are saved in rds files}
\usage{
NMcompAllRes(dir, ..., debug = F)
}
\arguments{
\item{dir}{The directory in which to look for nonmem results}

\item{...}{passed to NMcompRes which will pass them to NMscanData}

\item{debug}{start by calling browser()?}
}
\description{
make sure all Nonmem results are saved in rds files
}
\seealso{
Other DataWrangling: \code{\link{NMaddColumns}},
  \code{\link{NMorderColumns}}, \code{\link{NMscanDataDF}},
  \code{\link{NMscanData}}, \code{\link{compareNames}},
  \code{\link{factor2}}, \code{\link{findCovs_df}},
  \code{\link{findCovs}}, \code{\link{findVars}},
  \code{\link{mergeCheck}}, \code{\link{newNames}},
  \code{\link{objInfo}}, \code{\link{rbindIntersect}},
  \code{\link{rbindUnion}}, \code{\link{roundDF}},
  \code{\link{roundResp0}}, \code{\link{selectVars}},
  \code{\link{signif2}}, \code{\link{stampObj}},
  \code{\link{tmpcol}}
}
\concept{DataWrangling}
