% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMplotFit.R
\name{NMplotFit}
\alias{NMplotFit}
\title{Plot individual observations together with individual and/or population predictions}
\usage{
NMplotFit(data, grp = "dose", par.time = "TIME",
  par.ntim = "NOMTIME", log.y = F, plot.obs = T, plot.ipre = T,
  plot.dvmean = F, plot.dvrange = F, plot.predmean = F, facet = T,
  type.mean, debug = F)
}
\arguments{
\item{data}{The dataset to plot}

\item{grp}{}

\item{par.ntim}{Used if plotting population means}
}
\description{
An summary plot of the observations and predictions
}
\seealso{
Other plotting: \code{\link{extractLegend}},
  \code{\link{theme_pp_facet}}, \code{\link{theme_pp}}
}
\concept{plotting}
