% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMplotFit.R
\name{NMplotFit}
\alias{NMplotFit}
\title{Plot individual observations together with individual and/or population predictions}
\usage{
NMplotFit(data, grp = "dose", par.time = "TIME",
  par.ntim = "NOMTIME", log.y = F, plot.obs = T, plot.ipre = T,
  plot.dvmean = F, plot.dvrange = F, plot.predmean = F, facet = T,
  type.mean, debug = F)
}
\arguments{
\item{data}{The dataset to plot.}

\item{grp}{character vector of columns to group means by and to facet plots by.}

\item{par.time}{name of column to plot observations against.}

\item{par.ntim}{Used if plotting population means.}

\item{log.y}{Use logarithmic scale for y-axis?}

\item{plot.obs}{Plot observations?}

\item{plot.ipre}{Plot individual predictions?}

\item{plot.dvmean}{Include mean of observations?}

\item{plot.dvrange}{Include range of observations?}

\item{plot.predmean}{Include mean of population predictions?}

\item{facet}{name of a column to facet by (facet_wrap used if this argument is given).}

\item{type.mean}{Passed to means. When linear scale is used, arithmetic is used, when on log scale, geometric is default.}

\item{debug}{Start by calling browser()?}
}
\description{
An summary plot of the observations and predictions.
}
\seealso{
Other plotting: \code{\link{extractLegend}},
  \code{\link{theme_pp_facet}}
}
\concept{plotting}
