% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMreadTab.R
\name{NMreadTab}
\alias{NMreadTab}
\title{Read a table file from NONMEM}
\usage{
NMreadTab(file, silent = F, skip, sep = " ", debug = F)
}
\arguments{
\item{file}{path to NONMEM table file}

\item{silent}{logical stating whether or not information is
printed.}

\item{skip}{Amount of lines skipes in the start of the file. Some
files have additional headers so if table looks wierd try
adding 1 to skip.}
}
\value{
Nonmem table as df.
}
\description{
Read a table file from NONMEM
}
\author{
Philip Delff
}
