% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMrunLog.R
\name{NMrunLog}
\alias{NMrunLog}
\title{get an overview of your Nonmem runs}
\usage{
NMrunLog(dir, runs, runs.omit, debug = F)
}
\arguments{
\item{dir}{The directory in which to search for runs}

\item{runs}{Runs to include in case you only want a specific subset.}

\item{runs.omit}{Runs to omit in case you want all but these.}

\item{debug}{start by running browser()?}
}
\description{
get an overview of your Nonmem runs
}
\seealso{
Other Nonmem: \code{\link{NMgetSection}},
  \code{\link{NMreadCsv}}, \code{\link{NMreadRun}},
  \code{\link{NMreadTab}}, \code{\link{NMreplacePart}},
  \code{\link{NMscanTables}}, \code{\link{NMtransInput}},
  \code{\link{NMundoReplace}}, \code{\link{NMwriteData}},
  \code{\link{extFormat}}, \code{\link{sumofork}}
}
\concept{Nonmem}
