% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMtransInput.R
\name{NMtransInput}
\alias{NMtransInput}
\title{read input data and translate names according to the $INPUT section}
\usage{
NMtransInput(file, debug = F)
}
\arguments{
\item{file}{a .lst or a .mod. No matter which you provide, the .mod is required and is the only one to be read.}

\item{debug}{start by running browser()?}
}
\description{
Based on a nonmem run, this function finds the input data and reads it. But it reads it like the nonmem run by applying DROP arguments and alternative naming of columns in the nonmem run.
}
\seealso{
Other Nonmem: \code{\link{NMgetSection}},
  \code{\link{NMreadCsv}}, \code{\link{NMreadRun}},
  \code{\link{NMreadTab}}, \code{\link{NMreplacePart}},
  \code{\link{NMrunLog}}, \code{\link{NMscanTables}},
  \code{\link{NMundoReplace}}, \code{\link{NMwriteData}},
  \code{\link{extFormat}}, \code{\link{sumofork}}
}
\concept{Nonmem}
