% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindUnion.R
\name{rbindUnion}
\alias{rbindUnion}
\title{rbind data.frames by filling in NA's}
\usage{
rbindUnion(...)
}
\arguments{
\item{...}{data frames to stack. Must be of class data.frame.}
}
\value{
a data.frame containing all data
}
\description{
This function is an implementation of the functionality as in
plyr::rbind.fill for not being dependent on that package. The
function is no longer maintained, and it is advised to use
data.table::rbind(...,fill) instead.
}
\details{
Consider using data.table::rbindlist(...,fill=TRUE) instead.
}
\examples{
x <- data.frame(a=1:3)
y <- data.frame(b=1:3,c="h")
z <- data.frame(a=4:5,b=letters[4:5],d=c(NA,5))
pmxtricks:::rbindUnion(x,y,z)
}
\seealso{
Other DataWrangling: 
\code{\link{newNames}()},
\code{\link{rbindIntersect}()},
\code{\link{roundDF}()},
\code{\link{roundResp0}()}
}
\concept{DataWrangling}
