% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVars.R
\name{selectVars}
\alias{selectVars}
\title{Select only wanted columns - If any of them are missing, tell the user
which ones.}
\usage{
selectVars(data, vars)
}
\arguments{
\item{data}{data.frame or data.table to select columns from.}

\item{vars}{columns to select. A vector of characters.}
}
\description{
This is just a wrapper to data[,vars]. But first it checks if vars are in
data. And if not it is reported in an error to the user. This saves you the
manual investigation every time data[,vars] gives you the annoyingly
non-informative error "undefined columns selected".
}
\author{
Philip Delff
}
