% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif2.R
\name{signif2}
\alias{signif2}
\title{round to fixed number of significant digits}
\usage{
signif2(x, digits = 1, add = T, debug = F)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{number of significant digits to round to.}

\item{add}{pad with zeros where digits>nchar(x[i]). Only TRUE
supported.}

\item{debug}{start by calling browser.}
}
\description{
round to fixed number of significant digits
}
\examples{
x <- c(1.24e-4,1.1334e6,1.1,22.00000,10.00,1)
signif(x,3)
as.character(signif(x,3))
signif.c(x,3)
signif.c(x,3,add=T)
signif.c(c(.2,11.84),2)
}
