% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofork.R
\name{sumofork}
\alias{sumofork}
\title{Platform independent sumo}
\usage{
sumofork(model, use.model.path = TRUE, tableType = 2,
  format.estimate = "\% -#6.4g", format.rse = "\%#6.3g")
}
\description{
Platform independent sumo
}
\details{
This is a fork of nonmem2R::sumo with very little
    modification. Many thanks to Magnus for a great job!
}
