% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pp.R
\name{theme_pp}
\alias{theme_pp}
\title{A custom theme - aim is powerpoint}
\usage{
theme_pp(theme0 = ggplot2::theme_bw, ...)
}
\arguments{
\item{theme0}{The theme to modify}

\item{...}{Arguments passed to theme() after default values have
been set. This means ... can be used to overwrite the
defaults.}
}
\description{
A custom theme - aim is powerpoint
}
\seealso{
Other Plotting: \code{\link{ggIndProfs}},
  \code{\link{ggstamp}}
}
\concept{Plotting}
